functions/requirements.txt
firebase_functions~=0.1.0
Jinja2
requests
python-dotenv